limit the number of files in /source/init
- move navigation stuff to its own folder
- move API to its own file
- make more use of constants.js
