+++ source/lib/api.js
@@ -15,6 +15,3 @@ export const API = (pth: string, query: ?Object = null) => {
 	}
 	return url
 }
-
-export const GH_BASE_URL = 'https://github.com/StoDevX/AAO-React-Native'
-export const GH_NEW_ISSUE_URL = `${GH_BASE_URL}/issues/new`
