+++ source/lib/user-agent.js
