+++ source/navigation/constants.js
@@ -1,0 +1,5 @@
+// @flow
+
+import {IS_PRODUCTION} from '../lib/constants'
+
+export const persistenceKey = IS_PRODUCTION ? 'NavState' : null
