+++ source/root.js
@@ -3,10 +3,6 @@
 import {AppRegistry, YellowBox} from 'react-native'
 import App from './app'
 
-// I'm not importing the exported variable because I just want to initialize
-// the file here.
-import './init/bugsnag'
-
 YellowBox.ignoreWarnings([
 	// TODO: remove me after upgrading to RN 0.56
 	'Warning: isMounted(...) is deprecated',
